refactor: extract background into BackgroundCanvas and refactor ClientCanvas to use UserImageCanvas
